@@ -1,6 +1,6 @@
 <?php
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Drupal\hel_tpm_group;
 
@@ -27,7 +27,7 @@ final class GroupsWithoutAdmins {
    */
   public function __construct(
     private readonly EntityTypeManagerInterface $entityTypeManager,
-    private readonly Connection $database
+    private readonly Connection $database,
   ) {}
 
   /**
@@ -61,7 +61,7 @@ final class GroupsWithoutAdmins {
    * @throws \Drupal\Component\Plugin\Exception\PluginNotFoundException
    */
   private function groupHasAdminUsers($gid, $roles) {
-    $storage = \Drupal::entityTypeManager()->getStorage('group_content');
+    $storage = $this->entityTypeManager->getStorage('group_content');
 
     $query = $storage->getQuery()
       ->accessCheck(FALSE)
