@@ -1,18 +1,17 @@
 <?php
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Drupal\Tests\hel_tpm_group\Kernel;
 
 use Drupal\Core\Entity\EntityInterface;
+use Drupal\Tests\field\Traits\EntityReferenceFieldCreationTrait;
+use Drupal\Tests\group\Kernel\GroupKernelTestBase;
 use Drupal\group\Entity\Group;
 use Drupal\group\Entity\GroupInterface;
 use Drupal\group\PermissionScopeInterface;
 use Drupal\node\Entity\Node;
 use Drupal\node\Entity\NodeType;
-use Drupal\Tests\field\Traits\EntityReferenceFieldCreationTrait;
-use Drupal\Tests\field\Traits\EntityReferenceTestTrait;
-use Drupal\Tests\group\Kernel\GroupKernelTestBase;
 use Drupal\user\Entity\User;
 use Drupal\user\RoleInterface;
 
