@@ -3,6 +3,9 @@
 namespace Drupal\hel_tpm_url_shortener\Plugin\Block;
 
 use Drupal\Core\Block\BlockBase;
+use Drupal\Core\Form\FormBuilderInterface;
+use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Provides url shortener link block.
@@ -13,13 +16,37 @@ use Drupal\Core\Block\BlockBase;
  *   category = "hel tpm url shortener"
  * )
  */
-class ShortenUrlLinkBlock extends BlockBase {
+class ShortenUrlLinkBlock extends BlockBase implements ContainerFactoryPluginInterface {
+
+  /**
+   * Form builder.
+   *
+   * @var \Drupal\Core\Form\FormBuilderInterface
+   */
+  private FormBuilderInterface $formBuilder;
+
+  public function __construct(array $configuration, $plugin_id, $plugin_definition, FormBuilderInterface $form_builder) {
+    parent::__construct($configuration, $plugin_id, $plugin_definition);
+    $this->formBuilder = $form_builder;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
+    return new static(
+      $configuration,
+      $plugin_id,
+      $plugin_definition,
+      $container->get('form_builder')
+    );
+  }
 
   /**
    * {@inheritdoc}
    */
   public function build() {
-    return \Drupal::formBuilder()->getForm('Drupal\hel_tpm_url_shortener\Form\ShortenUrlLinkForm');
+    return $this->formBuilder->getForm('Drupal\hel_tpm_url_shortener\Form\ShortenUrlLinkForm');
   }
 
 }
