@@ -1,6 +1,6 @@
 <?php
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Drupal\hel_tpm_user_expiry\Form;
 
@@ -28,7 +28,7 @@ class UserExpirySettings extends FormBase {
       '#type' => 'checkbox',
       '#title' => $this->t('Enable user expiration'),
       '#default_value' => SettingsUtility::getUserExpirationStatus(),
-      '#description' => t("Unchecking this will prevent user expire cron job to add more users to expiry queue. The value is stored using the State API, so configuration updates will not affect it."),
+      '#description' => $this->t("Unchecking this will prevent user expire cron job to add more users to expiry queue. The value is stored using the State API, so configuration updates will not affect it."),
     ];
     $form['actions']['#type'] = 'actions';
     $form['actions']['submit'] = [
