@@ -4,11 +4,11 @@ namespace Drupal\Tests\service_manual_workflow\Kernel;
 
 use Drupal\Core\Access\AccessResultAllowed;
 use Drupal\Core\Access\AccessResultForbidden;
-use Drupal\group\PermissionScopeInterface;
 use Drupal\Tests\content_moderation\Traits\ContentModerationTestTrait;
 use Drupal\Tests\group\Kernel\GroupKernelTestBase;
 use Drupal\Tests\node\Traits\ContentTypeCreationTrait;
 use Drupal\Tests\node\Traits\NodeCreationTrait;
+use Drupal\group\PermissionScopeInterface;
 use Drupal\user\RoleInterface;
 use Drupal\workflows\Entity\Workflow;
 use Prophecy\PhpUnit\ProphecyTrait;
