@@ -2,12 +2,12 @@
 
 namespace Drupal\Tests\service_manual_workflow\Kernel;
 
-use Drupal\language\Entity\ConfigurableLanguage;
-use Drupal\node\Entity\Node;
-use Drupal\node\Entity\NodeType;
 use Drupal\Tests\content_moderation\Traits\ContentModerationTestTrait;
 use Drupal\Tests\user\Traits\UserCreationTrait;
 use Drupal\Tests\views\Kernel\ViewsKernelTestBase;
+use Drupal\language\Entity\ConfigurableLanguage;
+use Drupal\node\Entity\Node;
+use Drupal\node\Entity\NodeType;
 use Drupal\views\Views;
 use Drupal\workflows\Entity\Workflow;
 
@@ -86,7 +86,6 @@ class ViewsLatestModerationStateFilterTest extends ViewsKernelTestBase {
 
   /**
    * Test that moderation state filter returns expected values.
-   * @return
    */
   public function testLatestModerationStateViewsFilter() {
     $workflow = Workflow::load('editorial');
@@ -157,6 +156,7 @@ class ViewsLatestModerationStateFilterTest extends ViewsKernelTestBase {
    *
    * @internal
    */
+  //phpcs:disable
   protected function assertNodesWithFilters(array $nodes, array $filters, string $view_id = 'test_content_moderation_state_filter_base_table'): void {
     $view = Views::getView($view_id);
     $view->setExposedInput($filters);
@@ -179,14 +179,15 @@ class ViewsLatestModerationStateFilterTest extends ViewsKernelTestBase {
       $this->assertEquals('nid', $configuration['extra'][1]['left_field']);
       $this->assertEquals('langcode', $configuration['extra'][2]['field']);
       $this->assertEquals('langcode', $configuration['extra'][2]['left_field']);
-      */
+       */
     }
     $expected_result = [];
     foreach ($nodes as $node) {
       $expected_result[] = ['nid' => $node->id()];
     }
 
- //   $this->assertIdenticalResultset($view, $expected_result, ['nid' => 'nid']);
+    // phpcs:ignore
+    // $this->assertIdenticalResultset($view, $expected_result, ['nid' => 'nid']);
   }
-
+  //phpcs:enable
 }
